requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-cartesia

mem0ai
duckduckgo-search
langchain_community

requests
httpx
python-dotenv
apscheduler
dateparser
python-dateutil
python-kasa
gitpython
watchdog
openai